Reorganize code and allow exporting tree data to OBJ
CURA-12636
